+++ Assets/Abuilities/AbilityItemProto.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6e6797f699fc7794da4301015cede600
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
