+++ Assets/Abuilities/Silkshiled/ParryItem.cs
@@ -1,34 +1,17 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using Abuilities;
 using player.script;
 using SaveAndLoad;
 using UnityEngine;
 
-public class ParryItem : MonoBehaviour
+class ParryItem : AbilityItemProto
 {
-    public static bool consume;
-    private BoxCollider2D bxCollider2D;
-    private ParticleSystem particleSys;
-    private void Start()
+    protected override void SetAbil()
     {
-        particleSys = GetComponentInChildren<ParticleSystem>();
-        bxCollider2D = GetComponent<BoxCollider2D>();
-        consume = false;
-        particleSys.Play();
-    }
-
-    private void Update()
-    {
-        
-    }
-
-    private void OnTriggerEnter2D(Collider2D other)
-    {
-        SaveData.SetAbilities("parry");
-        particleSys.Stop();
-        consume = true;
-        Parry.unlockParry = true;
         PlayerMove.canmove = false;
+        Parry.unlockParry = true;
+        SaveData.SetAbilities("parry");
     }
 }
