+++ Assets/Abuilities/Silkshiled/TheOrbItem.cs
@@ -1,5 +1,6 @@
 
 using System;
+using Abuilities;
 using player.script;
 using SaveAndLoad;
 using UnityEngine;
@@ -7,11 +8,13 @@ using UnityEngine;
 public class TheOrb : MonoBehaviour
 {
     private GameObject parentObj;
+    private AbilityItemProto abilityItem;
     [SerializeField] private GameObject descriptionText;
 
     private void Start()
     {
         parentObj = transform.parent.gameObject;
+        abilityItem = parentObj.GetComponent<AbilityItemProto>();
         if (Parry.unlockParry)
         {
             descriptionText.SetActive(false);
@@ -22,7 +25,7 @@ public class TheOrb : MonoBehaviour
     private void Update()
     {
         transform.eulerAngles += new Vector3(0f,0f,30f*Time.deltaTime);
-        if (!ParryItem.consume) return;
+        if (!abilityItem.consume) return;
             transform.position = Vector2.MoveTowards(transform.position, PlayerMove.playerPos, 0.8f * Time.deltaTime);
             if (new Vector2(transform.position.x, transform.position.y) != PlayerMove.playerPos) return;
             PlayerMove.canmove = true;
